removed the Signal change tracking feature
It was fundamentally wrong because it broke the property:

signal_lift!(s1, s2 => f) == s1.map(|a| s2.map(|b| f(a, b)))

The change tracking mechanism only considered the source signal as the
trigger for updating (running the closure). But in the lift case there are
other signals stored in the closure, so the current change tracking system
can't know their state and will never update them.