actually correct implementation of signal_lift
The nested map version was wrong because it produced a nested type
`Signal<Signal<...>>` but I went full retard and kept going because
it passed the tests and never looked at the return type of the macro.
The tests were wrong but they passed because at some point in the chain
it was calling `format!("({})", s)`, and there the result was being
recursively sampled by the formatter until it found the integer to print.