improved the signal_lift macro
The macro now does a let binding for each signal argument, so it behaves
more like a function. Before it would just paste the expression inside the
closure and cause weird behavior. Also added a special case for closure
arguments.

The closure case is needed to avoid producing expressions like:
(|a, b| a + b)(arg1, arg2)
that often cause type inference to fail.