don't use swap_remove to delete dead callbacks
Using swap_remove means that deleting stream observers has the side effect
of altering the order of the callbacks. Even if relying on this order is
a bad idea, it can cause unexpected and hard to debug problems on some
stream chains that intentionally or not rely on this order.