switched Storage from Mutex to RwLock
Mutex has the nice property that it gives Sync to any tipe and it's faster,
but it doesn't allow concurrent sampling, and that should be possible.
After this change we can't store non-Sync types in a signal, so the user
must wrap them (with something like Mutex or Fragile).
Most methods don't require `T: Send + Sync` so this isn't much of a deal
anymore, also the parking_lot RwLock implementation is pretty fast.
This also allows easily swapping the storage implementation to something
else like RCU or STM.